1. 增加用户权限设计
2. 修复一些bug

+++ src/backend/bisheng/api/v1/knowledge.py
@@ -1,4 +1,5 @@
 import asyncio
+import json
 import time
 from typing import List
 from uuid import uuid4
@@ -10,10 +11,13 @@ from bisheng.database.models.knowledge import (Knowledge, KnowledgeCreate,
                                                KnowledgeRead)
 from bisheng.database.models.knowledge_file import (KnowledgeFile,
                                                     KnowledgeFileRead)
+from bisheng.database.models.user import User
 from bisheng.interface.embeddings.custom import OpenAIProxyEmbedding
+from bisheng.settings import settings
 from bisheng.utils.logger import logger
 from fastapi import APIRouter, Depends, File, HTTPException, UploadFile
 from fastapi.encoders import jsonable_encoder
+from fastapi_jwt_auth import AuthJWT
 from langchain.embeddings.openai import OpenAIEmbeddings
 from langchain.vectorstores import Milvus
 from sqlmodel import Session, select
@@ -28,34 +32,50 @@ async def upload_file(*, file: UploadFile = File(...)):
         file_name = file.filename
         # 缓存本地
         file_path = save_uploaded_file(file.file, 'bisheng').as_posix()
-        return UploadFileResponse(file_path=file_path + '_' + file_name, )
+        return UploadFileResponse(file_path=file_path + '_' + file_name,)
+    except Exception as exc:
+        logger.error(f'Error saving file: {exc}')
+        raise HTTPException(status_code=500, detail=str(exc)) from exc
+
+
+@router.get('/embedding_param', status_code=201)
+async def get_embedding():
+    try:
+        # 获取本地配置的名字
+        model_list = settings.embedding_config
+        models = list(model_list.keys())
+        return {'data': {'models': models}}
     except Exception as exc:
         logger.error(f'Error saving file: {exc}')
         raise HTTPException(status_code=500, detail=str(exc)) from exc
 
 
 @router.post('/process', status_code=201)
-async def process_knowledge(
-    *, session: Session = Depends(get_session), data: dict
-):
+async def process_knowledge(*, session: Session = Depends(get_session), data: dict, Authorize: AuthJWT = Depends()):
     """上传文件到知识库.
     使用flowchain来处理embeding的流程
         """
+    Authorize.jwt_required()
+    payload = json.loads(Authorize.get_jwt_subject())
+
     knowledge_id = data.get('knowledge_id')
     chunck_size = data.get('chunck_size')
     file_path = data.get('file_path')
-    knowledge = session.exec(
-        select(Knowledge).where(Knowledge.id == knowledge_id)
-    ).one()
+
+    knowledge = session.exec(select(Knowledge).where(Knowledge.id == knowledge_id)).one()
+    if payload.get('role') != 'admin' and knowledge.user_id != payload.get('user_id'):
+        raise HTTPException(status_code=500, detail='没有权限操作当前知识库')
     collection_name = knowledge.collection_name
     files = []
     file_paths = []
     for path in file_path:
         filepath, file_name = path.split('_', 1)
         md5_ = filepath.rsplit('/', 1)[1]
-        db_file = KnowledgeFile(
-            knowledge_id=knowledge_id, file_name=file_name, status=1, md5=md5_
-        )
+        db_file = KnowledgeFile(knowledge_id=knowledge_id,
+                                file_name=file_name,
+                                status=1,
+                                md5=md5_,
+                                user_id=payload.get('user_id'))
         session.add(db_file)
         session.commit()
         session.refresh(db_file)
@@ -63,13 +83,7 @@ async def process_knowledge(
         file_paths.append(filepath)
         logger.info(f'fileName={file_name} col={collection_name}')
     asyncio.create_task(
-        addEmbedding(
-            collection_name=collection_name,
-            chunk_size=chunck_size,
-            file_paths=file_paths,
-            knowledge_files=files
-        )
-    )
+        addEmbedding(collection_name=collection_name, chunk_size=chunck_size, file_paths=file_paths, knowledge_files=files))
 
     knowledge.update_time = db_file.create_time
     session.add(knowledge)
@@ -78,19 +92,18 @@ async def process_knowledge(
 
 
 @router.post('/create', response_model=KnowledgeRead, status_code=201)
-def create_knowledge(
-    *, session: Session = Depends(get_session), knowledge: KnowledgeCreate
-):
+def create_knowledge(*, session: Session = Depends(get_session), knowledge: KnowledgeCreate, Authorize: AuthJWT = Depends()):
+    Authorize.jwt_required()
+    payload = json.loads(Authorize.get_jwt_subject())
     """创建知识库."""
     db_knowldge = Knowledge.from_orm(knowledge)
-    know = session.exec(
-        select(Knowledge).where(Knowledge.name == knowledge.name)
-    ).all()
+    know = session.exec(select(Knowledge).where(Knowledge.name == knowledge.name)).all()
     if know:
         raise HTTPException(status_code=500, detail='name 重复')
     if not db_knowldge.collection_name:
         # 默认collectionName
         db_knowldge.collection_name = f'col_{int(time.time())}_{str(uuid4())[:8]}'
+    db_knowldge.user_id = payload.get('user_id')
     session.add(db_knowldge)
     session.commit()
     session.refresh(db_knowldge)
@@ -98,73 +111,73 @@ def create_knowledge(
 
 
 @router.get('/', response_model=List[KnowledgeRead], status_code=200)
-def get_knowledge(*, session: Session = Depends(get_session)):
+def get_knowledge(*, session: Session = Depends(get_session), Authorize: AuthJWT = Depends()):
+    Authorize.jwt_required()
+    payload = json.loads(Authorize.get_jwt_subject())
     """ 读取所有知识库信息. """
     try:
-        knowledges = session.exec(
-            select(Knowledge).order_by(Knowledge.update_time.desc())
-        ).all()
+        if 'admin' != payload.get('role'):
+            knowledges = session.exec(
+                select(Knowledge).where(Knowledge.user_id == payload.get('user_id')).order_by(
+                    Knowledge.update_time.desc())).all()
+        else:
+            knowledges = session.exec(select(Knowledge).order_by(Knowledge.update_time.desc())).all()
+        res = [jsonable_encoder(flow) for flow in knowledges]
+        if knowledges:
+            db_user_ids = {flow.user_id for flow in knowledges}
+            db_user = session.exec(select(User).where(User.user_id.in_(db_user_ids))).all()
+            userMap = {user.user_id: user.user_name for user in db_user}
+            for r in res:
+                r['user_name'] = userMap[r['user_id']]
+        return res
+
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e)) from e
-    return [jsonable_encoder(knowledge) for knowledge in knowledges]
 
 
-@router.get(
-    '/file_list/{knowledge_id}',
-    response_model=List[KnowledgeFileRead],
-    status_code=200
-)
-def get_filelist(
-    *,
-    session: Session = Depends(get_session),
-    knowledge_id: int,
-    page_size: int = 10,
-    page_num: int = 1
-):
+@router.get('/file_list/{knowledge_id}', response_model=List[KnowledgeFileRead], status_code=200)
+def get_filelist(*, session: Session = Depends(get_session), knowledge_id: int, page_size: int = 10, page_num: int = 1):
     """ 删除知识库信息. """
     knowledge = session.get(Knowledge, knowledge_id)
     if not knowledge:
         raise HTTPException(status_code=404, detail='knowledge not found')
     # 查找上传的文件信息
     files = session.exec(
-        select(KnowledgeFile).where(
-            KnowledgeFile.knowledge_id == knowledge_id
-        ).order_by(KnowledgeFile.update_time.desc()
-                   ).offset(page_size * (page_num-1)).limit(page_size)
-    ).all()
+        select(KnowledgeFile).where(KnowledgeFile.knowledge_id == knowledge_id).order_by(
+            KnowledgeFile.update_time.desc()).offset(page_size * (page_num - 1)).limit(page_size)).all()
     return [jsonable_encoder(knowledgefile) for knowledgefile in files]
 
 
 @router.delete('/{knowledge_id}', status_code=200)
-def delete_knowledge(
-    *, session: Session = Depends(get_session), knowledge_id: int
-):
+def delete_knowledge(*, session: Session = Depends(get_session), knowledge_id: int, Authorize: AuthJWT = Depends()):
+    Authorize.jwt_required()
+    payload = json.loads(Authorize.get_jwt_subject())
     """ 删除知识库信息. """
     knowledge = session.get(Knowledge, knowledge_id)
     if not knowledge:
         raise HTTPException(status_code=404, detail='knowledge not found')
+    if 'admin' != payload.get('role') and knowledge.user_id != payload.get('user_id'):
+        raise HTTPException(status_code=404, detail='没有权限执行操作')
     session.delete(knowledge)
     session.commit()
     return {'message': 'knowledge deleted successfully'}
 
 
 @router.delete('/file/{file_id}', status_code=200)
-def delete_knowledge_file(
-    *, session: Session = Depends(get_session), file_id: int
-):
+def delete_knowledge_file(*, session: Session = Depends(get_session), file_id: int, Authorize: AuthJWT = Depends()):
+    Authorize.jwt_required()
+    payload = json.loads(Authorize.get_jwt_subject())
     """ 删除知识文件信息 """
     knowledge_file = session.get(KnowledgeFile, file_id)
     if not knowledge_file:
         raise HTTPException(status_code=404, detail='文件不存在')
+    if 'admin' != payload.get('role') and knowledge_file.user_id != payload.get('user_id'):
+        raise HTTPException(status_code=404, detail='没有权限执行操作')
     knowledge = session.get(Knowledge, knowledge_file.knowledge_id)
     # 处理vectordb
     collection_name = knowledge.collection_name
     embeddings = OpenAIEmbeddings()
-    milvus = Milvus(
-        embedding_function=embeddings,
-        collection_name=collection_name,
-        connection_args=connection_args
-    )
+    milvus = Milvus(embedding_function=embeddings, collection_name=collection_name, connection_args=connection_args)
     pk = milvus.col.query(expr=f'file_id == {file_id}', output_fields=['pk'])
     res = milvus.col.delete(f"pk in {[p['pk'] for p in pk]}")
     logger.info(f'act=delete_vector file_id={file_id} res={res}')
@@ -173,37 +186,21 @@ def delete_knowledge_file(
     return {'message': 'knowledge file deleted successfully'}
 
 
-connection_args = {
-    'host': '192.168.106.116',
-    'port': '19530',
-    'user': '',
-    'password': '',
-    'secure': False
-}
+connection_args = {'host': '192.168.106.116', 'port': '19530', 'user': '', 'password': '', 'secure': False}
 
 
-async def addEmbedding(
-    collection_name, chunk_size: int, file_paths: List[str],
-    knowledge_files: List[KnowledgeFile]
-):
+async def addEmbedding(collection_name, chunk_size: int, file_paths: List[str], knowledge_files: List[KnowledgeFile]):
     embeddings = OpenAIProxyEmbedding()
     for index, path in enumerate(file_paths):
         knowledge_file = knowledge_files[index]
         try:
-            texts, metadatas = _read_chunk_text(
-                path, knowledge_file.file_name, chunk_size
-            )
-            [
-                metadata.update({'file_id': knowledge_file.id})
-                for metadata in metadatas
-            ]
-            Milvus.from_texts(
-                texts=texts,
-                embedding=embeddings,
-                metadatas=metadatas,
-                collection_name=collection_name,
-                connection_args=connection_args
-            )
+            texts, metadatas = _read_chunk_text(path, knowledge_file.file_name, chunk_size)
+            [metadata.update({'file_id': knowledge_file.id}) for metadata in metadatas]
+            Milvus.from_texts(texts=texts,
+                              embedding=embeddings,
+                              metadatas=metadatas,
+                              collection_name=collection_name,
+                              connection_args=connection_args)
 
             session = next(get_session())
             db_file = session.get(KnowledgeFile, knowledge_file.id)
@@ -220,16 +217,9 @@ async def addEmbedding(
 
 
 def _read_chunk_text(input_file, file_name, size):
-    from langchain.document_loaders import (
-        PyPDFLoader, BSHTMLLoader, TextLoader, UnstructuredMarkdownLoader
-    )
+    from langchain.document_loaders import (PyPDFLoader, BSHTMLLoader, TextLoader, UnstructuredMarkdownLoader)
     from langchain.text_splitter import CharacterTextSplitter
-    filetype_load_map = {
-        'txt': TextLoader,
-        'pdf': PyPDFLoader,
-        'html': BSHTMLLoader,
-        'md': UnstructuredMarkdownLoader
-    }
+    filetype_load_map = {'txt': TextLoader, 'pdf': PyPDFLoader, 'html': BSHTMLLoader, 'md': UnstructuredMarkdownLoader}
 
     file_type = file_name.split('.')[-1]
     if file_type not in filetype_load_map:
@@ -237,9 +227,7 @@ def _read_chunk_text(input_file, file_name, size):
 
     loader = filetype_load_map[file_type](input_file)
     documents = loader.load()
-    text_splitter = CharacterTextSplitter(
-        chunk_size=size, chunk_overlap=0, add_start_index=True
-    )
+    text_splitter = CharacterTextSplitter(chunk_size=size, chunk_overlap=0, add_start_index=True)
     texts = text_splitter.split_documents(documents)
     raw_texts = [t.page_content for t in texts]
     metadatas = []
